requirements/default.txt
beautifulsoup4
fastapi
python-jose[cryptography]
python-multipart
requests
uvicorn[standard]
